[Rust] Initial filesystem walker & cache
When operating on a slow disk like a network Samba mounted over wifi, it can
take about 5 minutes to list approximately 50,000 mp3 files. Since our files
don't often change, we can cache the results of this file scan locally and
reload it from cache on startup.

Dependencies:
- We use walkdir to walk through the dir and get all the file listings. It is
  several times faster than just doing it through `fs:read_dir()`.
- We are using 'clap' for command line arg parsing.
- We are using "sled" as the DB, with each entry being an integer key mapped to
  an encoded string of the Path.
- We use bincode to encode strings for sled.

In the future I expect to be loading these files and transforming them into
an in-memory structure. I think it's sensible to cache them close to their
raw form as directory listings because we can expect to change the in-memory
form frequently, so keeping the cache simple seems like a good idea.